Refactor file link sidebar to use form in panel overlay.
